requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
pipreqs==0.4.11
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
yarg==0.1.9
